+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - 完善 `siyuan-sdk` 介绍 | Improve the introduction of `siyuan-sdk`
 - 修复拼写错误 | Fix spelling errors
 - 更新 `siyuan-sdk` 版本 | Update `siyuan-sdk` version
+- 添加 `块类型` 与 `块属性` 说明 | Add `Block Type` and `Block Attribute` descriptions